refactor: move handleXml to common
