+++ libs/vre/resource-editor/resource-properties/src/lib/resource-payloads-mapping.ts
@@ -47,9 +47,9 @@ import {
 } from '@dasch-swiss/dsp-js';
 import { CalendarDate, CalendarPeriod, GregorianCalendarDate } from '@dasch-swiss/jdnconvertiblecalendar';
 import { CustomRegex } from '@dasch-swiss/vre/shared/app-common';
+import { handleXML } from '../../../../shared/app-common/src/lib/handle-xml';
 import { DateTime } from './date-time';
 import { convertTimestampToDateTime, dateTimeTimestamp } from './date-time-timestamp';
-import { handleXML } from './handle-xml';
 import { populateValue } from './populate-value-method';
 
 interface MappingParameters<T extends ReadValue> {
