+++ libs/vre/shared/app-common/src/index.ts
@@ -7,3 +7,4 @@ export * from './lib/directives/mat-autocomplete-options-scroll.directive';
 export * from './lib/form-validators/at-least-one-string-required.validator';
 export * from './lib/animations/replace-animation';
 export * from './lib/rxjs-operators/filter-undefined-operator';
+export * from './lib/handle-xml';
