+++ libs/vre/shared/app-common/src/lib/handle-xml.ts
